# Conflicts:
#	app/src/UI/Features/LegacyMap/helpers/components/Coordinates/Coordinates.tsx
#	app/src/UI/Features/LegacyMap/helpers/functional/recordset-layers.ts

+++ app/src/UI/Features/LegacyMap/helpers/components/DrawControls.tsx
@@ -18,11 +18,15 @@ import {
 } from 'UI/Features/LegacyMap/helpers/functional/geo-tracking-mode';
 import { WhatsHereBoxMode } from 'UI/Features/LegacyMap/helpers/functional/whats-here-box-mode';
 import GeoShapes from 'constants/geoShapes';
-import TargetMode from 'constants/targetModes';
-import { GEO_TRACKING_FEATURE } from 'UI/Features/LegacyMap/helpers/functional/constants';
+import GeoTracking from 'state/actions/geotracking/GeoTracking';
+import { TargetMode } from 'constants/targetModes';
+import { GEO_TRACKING_FEATURE, SUBMITTED_ACTIVITY_SHAPE } from 'UI/Features/LegacyMap/helpers/functional/constants';
+
 import { DrawModeDisplay, EditControls } from 'UI/Features/LegacyMap/helpers/components/MapCustomControls';
 import Alerts from 'state/actions/alerts/Alerts';
 import mappingAlertMessages from 'constants/alerts/mappingAlerts';
+import { isDrawing, isPaused, isTracking } from 'utils/geoTrackingHelpers';
+import { GeoTrackingStatus } from 'constants/geoTrackingStatus';
 
 // @ts-expect-error mapboxdraw compatibility with maplibre-gl issue
 MapboxDraw.constants.classes.CONTROL_BASE = 'maplibregl-ctrl';
@@ -37,25 +41,38 @@ const DrawControls = () => {
   const whatsHereToggle = useSelector((state) => state.Map.whatsHere.toggle);
   const tileCacheMode = useSelector((state) => state.Map.tileCacheMode);
   const drawingCustomLayer = useSelector((state) => state.Map.drawingCustomLayer);
-  const appModeURL = useSelector((state) => state.AppMode.url);
-  const currGeoTrackingMode = useSelector((state) => state.Map.track_me_draw_geo.isTracking);
+
+  const geoTrackingStatus = useSelector((state) => state.Map.track_me_draw_geo.status);
+  const currGeoTrackingMode = isTracking(geoTrackingStatus);
+  const isDrawingShape = isDrawing(geoTrackingStatus);
+  const isPausedDrawing = isPaused(geoTrackingStatus);
   const [prevGeoTrackingMode, setPrevGeoTrackingMode] = useState<boolean>(false);
+
   const EMPTY_OBJECT = {}; //  a stable reference for the default value to avoid unnecessary re-renders
   const activityGeo = (useSelector((state) => state.ActivityPage.activity?.geometry) ?? [])[0] ?? EMPTY_OBJECT;
 
+  const can_edit = useSelector((state) => !!state.ActivityPage?.activeActivityPermissions?.can_edit);
+  const created_by = useSelector((state) => state.ActivityPage?.activity?.created_by);
+  const username = useSelector((state) => state.Auth.username);
+  const userCanEdit = username === created_by || can_edit;
+
+  const url = useSelector((state) => state.AppMode.url);
+
   const dispatch = useDispatch();
+  const uHistory = useHistory();
+
   const drawInstance = useRef<MapboxDraw>();
   const drawModeDisplay = useRef<DrawModeDisplay>();
   const editControls = useRef<EditControls>();
   const isEditing = useRef(false);
 
-  const uHistory = useHistory();
-
-  const [mode, setMode] = useState<TargetMode>(TargetMode.DISABLED);
-  const isEditDisabled = ![TargetMode.ACTIVITY].includes(mode);
-
   // keep a ref to mode so we don't need to keep re-binding the callback for maplibre. keep it in sync with a hook.
   const modeRef = useRef<TargetMode>(TargetMode.DISABLED);
+
+  const [mode, setMode] = useState<TargetMode>(TargetMode.DISABLED);
+  const prevMode = useRef<TargetMode>(TargetMode.DISABLED);
+
+  const isEditDisabled = ![TargetMode.ACTIVITY].includes(mode);
 
   const handleEdit = () => {
     const features = drawInstance.current?.getAll().features;
@@ -65,6 +82,11 @@ const DrawControls = () => {
     isEditing.current = true;
     drawInstance?.current?.changeMode('direct_select', { featureId: features[0].id });
     dispatch(Alerts.create(mappingAlertMessages.saveActivityShape));
+
+    if (prevMode.current === TargetMode.ACTIVITY_GEO_TRACK) {
+      // to disable geo-tracking resume button
+      dispatch(GeoTracking.edit(true));
+    }
   };
 
   const handleSave = () => {
@@ -75,11 +97,24 @@ const DrawControls = () => {
     if (!updatedFeature || updatedFeature.geometry.type === GeoShapes.Point) return;
 
     dispatch({ type: MAP_ON_SHAPE_UPDATE, payload: updatedFeature });
+
+    if (prevMode.current === TargetMode.ACTIVITY_GEO_TRACK) {
+      dispatch(GeoTracking.edit(false));
+    }
   };
 
   const hasEditableShape = () => {
     const features = drawInstance.current?.getAll().features;
-    return features && features.length > 0 && features[0].geometry.type !== GeoShapes.Point;
+
+    if (!features || features.length === 0 || !userCanEdit) return false;
+
+    const isGeoTracking = mode === TargetMode.ACTIVITY_GEO_TRACK;
+
+    if (isGeoTracking) {
+      return isPausedDrawing || geoTrackingStatus === GeoTrackingStatus.COMPLETED;
+    }
+
+    return features[0].geometry.type !== GeoShapes.Point;
   };
 
   const updateEditControlState = () => {
@@ -93,6 +128,41 @@ const DrawControls = () => {
     const shouldEnableEdit = hasEditableShape() && !isEditDisabled;
     editControls.current?.setDisabled(!shouldEnableEdit);
   }, [mode]);
+
+  useEffect(() => {
+    updateEditControlState();
+  }, [isDrawingShape, userCanEdit, activityGeo?.geometry]);
+
+  // make a geomtery for perviously submitted shapes
+  useEffect(() => {
+    if (!activityGeo || mode === TargetMode.ACTIVITY_GEO_TRACK) return;
+
+    const feature = drawInstance?.current?.getAll()?.features?.[0];
+    const submittedShapeFeature = drawInstance?.current?.get(SUBMITTED_ACTIVITY_SHAPE);
+    const isActivityGeoEmpty = (activityGeo?.geometry?.coordinates?.length ?? 0) === 0;
+    const isFeaturePresent = (feature?.geometry?.coordinates?.length ?? 0) > 0;
+
+    if (submittedShapeFeature?.id && !userCanEdit) {
+      drawInstance?.current?.delete(String(submittedShapeFeature.id));
+    }
+
+    // Early return if:
+    // - No geometry to draw
+    // - A feature is already present
+    // - User is not the creator or cannot edit
+    // - Not on the Activity page
+    if (isActivityGeoEmpty || isFeaturePresent || !userCanEdit || !url?.includes('Activity')) return;
+
+    drawInstance?.current?.deleteAll();
+    drawInstance?.current?.add({
+      id: SUBMITTED_ACTIVITY_SHAPE,
+      type: 'Feature',
+      geometry: activityGeo.geometry,
+      properties: {}
+    });
+
+    updateEditControlState();
+  }, [activityGeo?.geometry, userCanEdit, url]);
 
   // update drawn LineString or Polygon to a red dotted line if an error occurs
   useEffect(() => {
@@ -126,20 +196,9 @@ const DrawControls = () => {
   useEffect(() => {
     const feature = drawInstance?.current?.get(GEO_TRACKING_FEATURE);
     const isActivityGeoEmpty = Object.keys(activityGeo).length === 0;
-    const isFeaturePresent = feature && Object.keys(feature).length > 0;
-
-    if (isActivityGeoEmpty && isFeaturePresent) {
-      drawInstance?.current?.deleteAll();
-      return;
-    }
-
-    if (!activityGeo?.geometry) return;
-
-    const isPolygon = activityGeo.geometry.type === GeoShapes.Polygon;
-    const coordinates = activityGeo.geometry.coordinates;
+    const isFeaturePresent = feature && feature?.geometry?.coordinates?.length > 0;
+    const coordinates = activityGeo?.geometry?.coordinates || [];
     const hasError = String(activityGeo?.properties?.error ?? 'false') === 'true';
-    const isInitialDraw = !feature || coordinates.length === 1;
-    const justExitedTracking = prevGeoTrackingMode && !currGeoTrackingMode;
 
     const handleInitialDraw = () => {
       setPrevGeoTrackingMode(currGeoTrackingMode);
@@ -157,6 +216,17 @@ const DrawControls = () => {
       setPrevGeoTrackingMode(currGeoTrackingMode);
     };
 
+    if (isActivityGeoEmpty && isFeaturePresent) {
+      drawInstance?.current?.deleteAll();
+      handleUpdate();
+      return;
+    }
+
+    if (!activityGeo?.geometry || !isFeaturePresent) return;
+    const isPolygon = activityGeo.geometry.type === GeoShapes.Polygon;
+
+    const isInitialDraw = !feature || coordinates.length === 1;
+    const exitedTrackingAndDrawing = !prevGeoTrackingMode;
     if (mode === TargetMode.ACTIVITY_GEO_TRACK) {
       if (isInitialDraw) {
         handleInitialDraw();
@@ -165,7 +235,7 @@ const DrawControls = () => {
       }
     }
 
-    if (justExitedTracking && isPolygon) {
+    if (exitedTrackingAndDrawing && isPolygon) {
       handlePolygonConversion();
     }
   }, [activityGeo?.geometry, currGeoTrackingMode]);
@@ -282,24 +352,22 @@ const DrawControls = () => {
         break;
       }
     }
-
-    updateEditControlState();
   }, []);
 
   // setup mode based on what's going on in the redux store / current url
   useEffect(() => {
     disableDrawButtons(false);
+
     if (whatsHereToggle) {
       setMode(TargetMode.WHATS_HERE);
       return;
     } else if (tileCacheMode) {
       setMode(TargetMode.TILE_CACHE);
-
       return;
     } else if (drawingCustomLayer) {
       setMode(TargetMode.CUSTOM_LAYER);
       return;
-    } else if (appModeURL?.includes('Activity')) {
+    } else if (url?.includes('Activity')) {
       if (currGeoTrackingMode || prevGeoTrackingMode) {
         setMode(TargetMode.ACTIVITY_GEO_TRACK);
         disableDrawButtons(true);
@@ -309,8 +377,9 @@ const DrawControls = () => {
     } else {
       setMode(TargetMode.DISABLED);
       disableDrawButtons(true);
+      editControls.current?.reset();
     }
-  }, [whatsHereToggle, tileCacheMode, drawingCustomLayer, appModeURL, currGeoTrackingMode, prevGeoTrackingMode]);
+  }, [whatsHereToggle, tileCacheMode, drawingCustomLayer, url, currGeoTrackingMode, prevGeoTrackingMode]);
 
   /**
    * @desc Update the Drawn Shape.
@@ -320,7 +389,6 @@ const DrawControls = () => {
    */
   const drawShapeUpdate = useCallback((event, map: InvasivesMap | undefined) => {
     if (!drawInstance.current) return;
-
     const currentMode = drawInstance.current.getMode();
 
     if (currentMode === 'direct_select') {
@@ -349,13 +417,16 @@ const DrawControls = () => {
     if (editedGeo?.id !== featureId) {
       dispatch({ type: MAP_ON_SHAPE_UPDATE, payload: editedGeo });
     }
-    updateEditControlState();
   }, []);
 
   useEffect(() => {
     if (!drawInstance.current) return;
 
-    drawInstance.current.deleteAll();
+    const shouldPreserveShape = prevMode.current === TargetMode.ACTIVITY_GEO_TRACK && mode === TargetMode.ACTIVITY;
+
+    if (!shouldPreserveShape) {
+      drawInstance.current.deleteAll();
+    }
 
     switch (mode) {
       case TargetMode.WHATS_HERE:
@@ -375,6 +446,7 @@ const DrawControls = () => {
     }
 
     drawModeDisplay.current?.setMode(mode);
+    prevMode.current = mode;
   }, [mode]);
 
   useEffect(() => {
